% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr.R
\name{asr}
\alias{asr}
\title{Fill the Annual Safety Report template}
\usage{
asr(
  data,
  target = "tmp.docx",
  trial_title = "TRIAL NAME",
  protocol_number = "default",
  basec_number = "default",
  snctp_number = "default",
  swissmedic_number = "default",
  ec_name = "default",
  tr_number = "default",
  product_name = "default",
  sponsor_contact = "default name, default number, default email",
  inst_name_address = "default name, default address",
  n_centers_t = "default",
  n_centers_p = "default",
  n_centers_c = "default",
  n_centers_o = "default",
  n_pat_t = "default",
  n_pat_e = 300,
  n_pat_c = "default",
  n_pat_p = "default",
  n_centers_t_ch = "default",
  n_centers_p_ch = "default",
  n_centers_c_ch = "default",
  n_centers_o_ch = "default",
  n_pat_t_ch = "default",
  n_pat_e_ch = 300,
  n_pat_c_ch = "default",
  n_pat_p_ch = "default",
  report_date = format(Sys.Date(), format = "\%d/\%m/\%Y"),
  period_from = as.Date("2020-11-02"),
  period_to = as.Date("2020-11-17"),
  template = system.file("extdata/clino_annual_safety_report_fm.docx", package =
    "SwissASR"),
  international = FALSE,
  trial_type = "imp",
  var_class = "class",
  var_sae_n = "sae_n",
  var_part_id = "record_id",
  var_age = "age",
  var_sex = "sex",
  var_country = "country",
  var_site = "site",
  var_sae = "sae",
  var_date_onset = "sae_date",
  var_trt = "trt",
  var_date_trt_start = "sae_trtstart",
  var_date_trt_stop = "sae_trtstop",
  var_outcome = "outcome",
  var_comment = "comment",
  var_relation = "related",
  var_expected = "expected",
  var_devdef = "devdef",
  var_devattr = "devattr",
  var_devint = "devint",
  var_safetymeasure = "safetymeasure",
  var_tx = NULL
)
}
\arguments{
\item{data}{dataframe containing safety data}

\item{target}{filename to save the report to}

\item{trial_title}{name of the trial}

\item{protocol_number}{trial code/protocol number}

\item{basec_number}{BASEC number}

\item{snctp_number}{SNCTP number}

\item{swissmedic_number}{Swissmedic number}

\item{ec_name}{EC name (Lead EC and/concerned EC)}

\item{tr_number}{Number for Transplantation Clinical Trials (FOPH number)}

\item{product_name}{product name or intervention}

\item{sponsor_contact}{contact details of sponsor(-investigator)}

\item{inst_name_address}{name and address of institute}

\item{n_centers_t}{total number of participating centres}

\item{n_centers_p}{planned number of participating centres}

\item{n_centers_c}{number of closed centres}

\item{n_centers_o}{number of open centres}

\item{n_pat_t}{target number of participants}

\item{n_pat_e}{number of enrolled participants}

\item{n_pat_c}{number of completed participants}

\item{n_pat_p}{number of prematurely terminated participants}

\item{n_centers_t_ch}{total number of participating centres in CH}

\item{n_centers_p_ch}{planned number of participating centres in CH}

\item{n_centers_c_ch}{number of closed centres in CH}

\item{n_centers_o_ch}{number of open centres in CH}

\item{n_pat_t_ch}{target number of participants in CH}

\item{n_pat_e_ch}{number of enrolled participants in CH}

\item{n_pat_c_ch}{number of completed participants in CH}

\item{n_pat_p_ch}{number of prematurely terminated participants in CH}

\item{report_date}{report date}

\item{period_from}{start of reporting period}

\item{period_to}{end of reporting period}

\item{template}{path to template file}

\item{international}{internation or national trial (logical)}

\item{trial_type}{one of \code{imp}, \code{medical device}, \code{other}. Abbreviations OK.}

\item{var_class}{variable containing SAE class. Options allowed are "SAE", "SADR", "SUSAR"}

\item{var_sae_n}{variable containing SAE ID}

\item{var_part_id}{variable containing participant ID}

\item{var_age}{variable containing participant age}

\item{var_sex}{variable containing participant sex}

\item{var_country}{variable containing participants country}

\item{var_site}{variable containing participants site}

\item{var_sae}{variable containing SAE type (description)}

\item{var_date_onset}{variable containing the date the SAE occured}

\item{var_trt}{variable containing the treatment used to alleviate the SAE}

\item{var_date_trt_start}{variable containing the date that the treatment started}

\item{var_date_trt_stop}{variable containing the date that the treatment ended}

\item{var_outcome}{variable containing the outcome of the SAE}

\item{var_comment}{variable containing any comment}

\item{var_relation}{variable containing the relationship to randomized intervention}

\item{var_expected}{variable saying whether the SAE was expected}

\item{var_devdef}{variable containing whether the SAE is a device deficiency}

\item{var_devattr}{variable containing whether the SAE is attributable to the device}

\item{var_devint}{variable containing whether the SAE is attributable to an intervention in the trial}

\item{var_safetymeasure}{variable containing whether the SAE required safety related measures}

\item{var_tx}{variable indicating the intervention group. If provided, this variable will be included in the line listing}
}
\value{
nothing in R, creates a docx file in the \code{target} location
}
\description{
This function fills out the SwissEthics annual safety report template with a
much as possible. Minor (formatting) changes will still be required after generation of the
report.
}
\details{
Use of the \code{var_tx} argument results in the indicated variable being included in the line listing.
}
\examples{

data(asr_sae)
file <- tempfile("asr", fileext = ".docx")
asr(asr_sae, file)

# # more usual use will require passing more information:
# asr(asr_sae, file,
#     # trial info
#     trial_title = "Example Trial Name",
#     protocol_number = "20221002130",
#     basec_number = "",
#     snctp_number = "202200458",
#     swissmedic_number = "....",
#     ec_name = "Kantonale Ethikskommision Bern",
#     tr_number = "",
#     product_name = "Drug name",
#     international = FALSE,
#     trial_type = "imp",
#     # Sponsor info
#     sponsor_contact = "Sponsor name, Sponsor phone number, Sponsor email",
#     inst_name_address = "Institute name, Institute address",
#     # site info
#     n_centers_t = 20,        # total number
#     n_centers_p = "default", # planned
#     n_centers_c = "default", # closed
#     n_centers_o = "default", # open
#     # participant info
#     n_pat_t = 1000,          # target
#     n_pat_e = 300,           # enrolled
#     n_pat_c = 0,             # complete
#     n_pat_p = 0,             # prematurely terminated
#     # report info
#     report_date = format(Sys.Date(), format = "\%d/\%m/\%Y"),
#     period_from = as.Date("2020-11-02"),
#     period_to = as.Date("2020-11-17"),
#     # variable mapping
#     var_class = "class",
#     var_sae_n = "sae_n", #sae ID
#     var_part_id = "record_id", #participant ID
#     var_age = "age",
#     var_sex ="sex",
#     var_country = "country",
#     var_site = "site",
#     var_sae = "sae",
#     var_date_onset = "sae_date",
#     var_trt = "trt",
#     var_date_trt_start = "sae_trtstart",
#     var_date_trt_stop = "sae_trtstop",
#     var_outcome = "outcome",
#     var_comment = "comment",
#     var_relation = "related",
#     var_expected = "expected",
#     var_safetymeasure = "safetymeasure"
#     )



}
