% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_dataprep.R
\name{asr_dataprep}
\alias{asr_dataprep}
\title{Prepare data for the \code{asr} function
This function allows the use of the \code{asr_safety_summary} function outside
of the normal \code{asr} function, which might be desirable to get an overview
of the information in a different format to MS Word.}
\usage{
asr_dataprep(
  data,
  trial_type = "imp",
  international = FALSE,
  period_from = NA,
  period_to = NA,
  var_class = "class",
  var_sae_n = "sae_n",
  var_part_id = "record_id",
  var_age = "age",
  var_sex = "sex",
  var_country = "country",
  var_site = "site",
  var_sae = "sae",
  var_date_onset = "sae_date",
  var_trt = "trt",
  var_date_trt_start = "sae_trtstart",
  var_date_trt_stop = "sae_trtstop",
  var_outcome = "outcome",
  var_comment = "comment",
  var_relation = "related",
  var_expected = "expected",
  var_devdef = "devdef",
  var_devattr = "devattr",
  var_devint = "devint",
  var_safetymeasure = "safetymeasure"
)
}
\arguments{
\item{data}{dataframe containing safety data}

\item{trial_type}{one of \code{imp}, \code{medical device}, \code{other}. Abbreviations OK.}

\item{international}{internation or national trial (logical)}

\item{period_from}{start of reporting period}

\item{period_to}{end of reporting period}

\item{var_class}{variable containing SAE class. Options allowed are "SAE", "SADR", "SUSAR"}

\item{var_sae_n}{variable containing SAE ID}

\item{var_part_id}{variable containing participant ID}

\item{var_age}{variable containing participant age}

\item{var_sex}{variable containing participant sex}

\item{var_country}{variable containing participants country}

\item{var_site}{variable containing participants site}

\item{var_sae}{variable containing SAE type (description)}

\item{var_date_onset}{variable containing the date the SAE occured}

\item{var_trt}{variable containing the treatment used to alleviate the SAE}

\item{var_date_trt_start}{variable containing the date that the treatment started}

\item{var_date_trt_stop}{variable containing the date that the treatment ended}

\item{var_outcome}{variable containing the outcome of the SAE}

\item{var_comment}{variable containing any comment}

\item{var_relation}{variable containing the relationship to randomized intervention}

\item{var_expected}{variable saying whether the SAE was expected}

\item{var_devdef}{variable containing whether the SAE is a device deficiency}

\item{var_devattr}{variable containing whether the SAE is attributable to the device}

\item{var_devint}{variable containing whether the SAE is attributable to an intervention in the trial}

\item{var_safetymeasure}{variable containing whether the SAE required safety related measures}
}
\value{
a list of 2 dataframes (\code{data} containing all data submitted to
the function and \code{period_data} which contains only the data for the relevant
period) with variables renamed to those expected by \code{asr_safety_summary}
}
\description{
Prepare data for the \code{asr} function
This function allows the use of the \code{asr_safety_summary} function outside
of the normal \code{asr} function, which might be desirable to get an overview
of the information in a different format to MS Word.
}
\examples{
data(asr_sae)
# IMP
asr_dataprep(asr_sae, period_from = as.Date("2020-10-10"), period_to = as.Date("2021-10-10"))
asr_dataprep(asr_sae, period_from = min(asr_sae$sae_date), period_to = Sys.Date())
}
