% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asr_safety_summary.R
\name{asr_safety_summary}
\alias{asr_safety_summary}
\title{Summarize safety data}
\usage{
asr_safety_summary(data, period_data, trial_type, n_pat_e)
}
\arguments{
\item{data}{SAE data}

\item{period_data}{SAE data restricted to a specific period}

\item{trial_type}{trial type (\code{imp}, \code{medical device} or \code{other})}

\item{n_pat_e}{Number of enrolled participants}
}
\value{
\describe{
\item{txt}{The sentences required for the report}
\item{tab}{The summary table required for the report}
\item{tab_map}{Mapping between the variable name in \code{tab} and a nicer label}
}
}
\description{
This function creates the text and summary table for the report.
}
\examples{
data(asr_sae)
# IMP
prepped <- asr_dataprep(asr_sae,
                        period_from = as.Date("2020-10-10"),
                        period_to = as.Date("2021-10-10"))
asr_safety_summary(data = prepped$data, period_data = prepped$period_data,
                   "imp", 60)

# medical devices
prepped <- asr_dataprep(asr_sae, period_from = as.Date("2020-10-10"),
                        period_to = as.Date("2021-10-10"), trial_type = "m")
summ <- asr_safety_summary(data = prepped$data,
                           period_data = prepped$period_data, "m", 60)

# other trial
asr_safety_summary(data = prepped$data, period_data = prepped$period_data, "o", 60)
}
